Rename `getLiftedConfig` → `getConfig` & `getConfig` → `getRawConfig`
The previous `getConfig` returned a `BootupErrors[Config]` which was
rarely used (since calling any of the `get` methods on it would be not
lifted). So `getLiftedConfig` was the one used. This is annoying, so
rename them to make `getConfig` the default and correct way of getting a
safe config object.